Feat: Create Records folder
- Records.js 와 Records.module.css를 포함하는 Records 폴더 생성
